PROYECTOS/SOLANA_SNIPER_BOT/requirements.txt
# CIPHER Pump.fun Sniper Bot - Dependencies
# Python 3.10+ required

# Solana
solana>=0.32.0
solders>=0.20.0

# HTTP & WebSocket
websockets>=12.0
httpx>=0.25.0
aiohttp>=3.9.0

# Database
aiosqlite>=0.19.0

# Utils
python-dotenv>=1.0.0

# Optional: Better logging
rich>=13.7.0
